Done with first version of wave demo scene

+++ XiangARUnity/Assets/VRLionFixing/Script/Main/Input/InputInterface.cs
@@ -11,7 +11,7 @@ namespace Hsinpa.Input
             get;
         }
 
-        int raycastLength
+        float raycastLength
         {
             get;
         }
