+++ XiangARUnity/Assets/VRLionFixing/Script/Main/Input/InputStandalone.cs
@@ -14,7 +14,7 @@ public class InputStandalone : InputInterface
 
     public Vector3 faceDir { get =>  (GetMouseWorldPos() - _camera.transform.position).normalized; }
 
-    public int raycastLength => 100;
+    public float raycastLength => 100;
 
     public bool GetMouse()
     {
